[Refactoring][WIP] Error handling
