+++ Cookbook.API/Program.cs
@@ -1,4 +1,4 @@
-using Cookbook.API.Controllers;
+using Cookbook.API.ErrorHandeling;
 using Cookbook.API.Mapping;
 using Cookbook.Application;
 using Cookbook.Repository.Repositories;
@@ -6,14 +6,13 @@ using Cookbook.Repository.Repositories;
 var builder = WebApplication.CreateBuilder(args);
 var config = builder.Configuration;
 
-builder.Services.AddControllers(options =>
-{
-    options.Filters.Add<JsonExceptionFilter>();
-});
+
 builder.Services.AddControllers(options =>
 {
     options.Filters.Add<ValidationActionFilterAttribute>();
+    options.Filters.Add<JsonExceptionFilter>();
 });
+
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
